Re-rename module '_pint' back to to 'pint' because that wasn't the issue earlier
